refactor project structure + rename files
