+++ src/pages/Menu.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 
-import Gallery from './gallery';
+import Gallery from '../components/Gallery';
 
 import menuPDF from '../menu.pdf';
